requirements.txt
streamlit==1.31.1
python-dotenv==1.0.1
pypdf==4.0.2

langchain==0.2.11
langchain-community==0.2.10
langchain-groq==0.1.2

faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0